Compiler warning fixes

+++ playerbot/strategy/values/CraftValue.h
@@ -44,9 +44,9 @@ namespace ai
             for (map<uint32, int>::iterator i = required.begin(); i != required.end(); ++i)
             {
                 uint32 item = i->first;
-                if (obtained[item] >= required[item] * count)
+                if (obtained[item] >= required[item] * (int)count)
                 {
-                    obtained[item] -= required[item] * count;
+                    obtained[item] -= required[item] * (int)count;
                 }
             }
         }
